Return web links for OWNERS files when listing code owners
When listing code owners the response contains the code owner config
files (e.g. OWNERS files) that have been inspected to gather the code
owners. Knowing the inspected code owner config files is important if
you want to verify that the correct code owners have been returned or if
you want to understand why a certain expected code owner was not
included, as in this case you want to look into these code owner config
files to see which code owners and rules they define. So far, we
returned the project, branch and path for the code owner config files,
which is sufficient to locate the code owner config files, but
assembling the links to the code owner config files from this
information is cumbersome and can take significant time, especially if
you need to inspect multiple code owner config files. To save this
effort we now return file web links for the code owner config files.
Having file web links available makes the lookup of code owner config
files much easier, as they contain the direct URLs to the code owner
config files.

Note that file web links are only available if they are configured.

CodeOwnerConfigFileJson is the class that writes the JSON for the
returned code owner config files. In this class we need to set the new
CodeOwnerConfigFileInfo.webLinks field. To create the file web links we
need the project, branch and path of the code owner config file, but
also the revision from which it was read. The revision was not available
in CodeOwnerConfigFileJson since it only got the CodeOwnerConfig.Key as
input which doesn't contain the revision. To make the revision available
we have to provide CodeOwnerConfig instead of CodeOwnerConfig.Key.

CodeOwnerConfigFileJson also gets imported code owner config files as
input which are formatted in the same way as the importing code owner
config file. To include web links for them too CodeOwnerConfigImport
needs to be extended to contain CodeOwnerConfig, which contains the
revision, instead of only CodeOwnerConfig.Key.

Since some imports can be unresolved (e.g. an import that references an
non-existing code owner config file) the imported CodeOwnerConfig in
CodeOwnerConfigImport is optional (i.e. it is not set for unresolved
imports). This means, if an import is unresolved no web links are
returned for it.

Due to CodeOwnerConfigImport needing the imported CodeOwnerConfig for
resolved imports now, we need to provide it in PathCodeOwners which is
responsible for resolving the imports and also in quite some tests.

Bug: Google b/345161989
Change-Id: If7efad1ffaf456a192ccea2e09d76b19d8f3b4fa
Signed-off-by: Edwin Kempin <ekempin@google.com>
Reviewed-on: https://gerrit-review.googlesource.com/c/plugins/code-owners/+/434337
Reviewed-by: Kamil Musin <kamilm@google.com>
Tested-by: Zuul <zuul-63@gerritcodereview-ci.iam.gserviceaccount.com>