Add a basic screen to invoke the Check Code Owner REST endpoint
It happens frequently that users have questions about code owners, e.g.
why a certain user is or isn't a code owner of a certain file. In this
case admins can use the Check Code Owner REST endpoint to get insights
into the code owner computation.

Invoking the Check Code Owner REST endpoint is a bit cumbersome though
and hence takes more time than it should:

* one needs to lookup the URL from the documentation and then assemble
  the URL manually
* identifiers in the URL (e.g. project names and files names) need to be
  URL-encoded
* to make the JSON output readable pretty-printing (pp=1) should be
  requested
* one needs to know how to call the REST endpoint with authentication
  (e.g. through curl or by opening the URL in the browser)

To simplify this we add a very basic screen with a form that allows to
call the Check Code Owner REST endpoint from the UI. This form takes
care of the pain-points above and also provides links to the
documentation.

The new screen is available at:
<GERRIT_URL>/x/code-owners/check-code-owner

Screenshot: https://i.imgur.com/L0nzcWa.png

The implementation of the screen followed the example of the service
user creation screen from the serviceuser plugin:
https://gerrit.googlesource.com/plugins/serviceuser/+/master/web/gr-serviceuser-create.ts

Bug: Google b/345161989
Change-Id: I25cff9f08ee3c8eb3a11dc4cd2d262fc8be1a6e3
Signed-off-by: Edwin Kempin <ekempin@google.com>