CheckCodeOwner: Return relevant code owner config files
To make debugging of code owner config files easier return all code
owner config files that are relevant for computing the code ownership
when checking a code owner.

Note, there is already a codeOwnerConfigFilePaths field. This field
contains the paths of the code owner config files that assign code
ownership to the given email for the specified path. This is different
from the relevant code owner config files which are all code owner
config files which are inspected. In a follow-up change we may drop the
codeOwnerConfigFilePaths field and instead include the information of
whether a code owner config file assigns code ownership to the given
email for the specified path into a new field inside
CodeOwnerConfigFileInfo.

Bug: Google b/345161989
Change-Id: Ib3a57feac6e9ea293eaba97640c4251fd8d558e7
Signed-off-by: Edwin Kempin <ekempin@google.com>