PathCodeOwners: Flatten the return type (step 3/3)
Add a AutoValue builder for PathCodeOwnersResult.

This allows us to get rid of local vars to collect the result data.

Having an AutoValue builder for PathCodeOwnersResult also allows us to
unify how results within PathCodeOwners are passed between methods. At
the moment the resolveImports method has a return value to return
resolved/unresolved imports and the messages, and in addition it gets a
CodeOwnerConfig.Builder passed in to write out relevant parts of the
imported CodeOwnerConfig. Using 2 different approaches to return results
is confusing. If we have an AutoValue builder for PathCodeOwnersResult
we can just use this builder to collect all the results.

Change-Id: I69c13959dc08517d9dbc8e144d7a28b0fe388582
Signed-off-by: Edwin Kempin <ekempin@google.com>