Check Code Owner screen: Show note to non-admins that debug logs are limited
Full debug logs are only shown to administrators (users with the
'Administrate Server' capability or the 'Check Code Owner' capability).
If a user is not an admin show them a note saying that the debug logs
are limited and that they should reach out to an admin if the
information they got is not enough.

Bug: Google b/345161989
Change-Id: Ia973e3df48b140e9ad300a7e10513ba626d6cc44
Signed-off-by: Edwin Kempin <ekempin@google.com>