Fix transitive import of global owners via per-file rule that ignores global owners
Example:

OWNERS:
per-file *.md=set noparent
per-file *.md=file:/foo/OWNERS

/foo/OWNERS:
file:/bar/OWNERS

/bar/OWNERS:
code-owner@example.com

For a file that matches the per-file rule in 'OWNERS' we would expect
that 'code-owner@example.com' is a code owner. 'code-owner@example.com'
is a global code owner in '/bar/OWNERS' but since it's imported for a
per-file rule it becomes a per-file code owner. However what happens is
that we wrongly treat 'code-owner@example.com' as a global code owner
and then it gets ignored due to the "set noparent" on the matching
per-file rule.

In PathCodeOwners.resolveImports we already had logic to treat global
code owners that are imported via a per-file rule as per-file code
owners, however this worked only for global code owners that were
directly imported by a per-file rule, but not for global code owners
that were transitively imported by a per-file rule.

This got broken by the refactoring that was done in change I6f7723c0a,
but we didn't notice the breakage because we only tested this case
without transitive imports (see
PathCodeOwnersTest.perFileRuleThatIgnoresGlobalCodeOwnersCanImportGlobalCodeOwnersFromOtherFile).

To fix this we need to know for an import if any previous import in the
import chain was a per-file import. If yes, we should treat the imported
global code owners as per-file code owners.

The check whether an import imports code owners as global code owners or
as per-file code owners is done by
PathCodeOwners.CodeOwnerImport#isGlobal. So far it was only looking at
the current import but not at the previous imports. By passing in the
previous import (if there is one) we can fix this.

Bug: Issue b/363000012
Change-Id: I7c872175fb198603f24a95819877ee1ff98651ec
Signed-off-by: Edwin Kempin <ekempin@google.com>