Check permissions of uploader when validating imports on submit
We do validate code owner config files on upload. One validation that is
performed for imports is that the project/branch of the imported code
owner config is visible to the uploader.

It's possible that this project/branch is not visible to everyone. By
referring to another project/branch in a code owner config file the
uploader reveals the existence of the project/branch to everyone who can
see the code owner config file.

If enabled, the validation of code owner config files is also done on
submit. At this point it's intended to do the exact same validation as
on upload. In particular this means that all visibility checks should be
done from the perspective of the uploader (and not for the submitter)
[1].

We already did the visibility checks for the code owners for the
uploader, but wrongly checked the visibility of projects/branches from
which code owner config files are imported for the submitter. This
change fixes this so that visibility checks for projects/branches from
which code owner config files are imported are also done for the
uploader.

[1] https://gerrit-review.googlesource.com/plugins/code-owners/Documentation/validation.html#:~:text=If%20enabled%2C%20on%20submit%20we%20repeat%20the%20exact%20same%20validation%20that%20was%20done%20on%20upload.%20This%20means%2C%20all%20visibility%20checks%20will%20be%20done%20from%20the%20perspective%20of%20the%20uploader.

Bug: Google b/365550280
Change-Id: I6b1cc65f90f1fce1d9fb2930d40103390b6d10f0
Signed-off-by: Edwin Kempin <ekempin@google.com>