Fix RepoView leak in CodeOwnerConfigValidator
The onPreMerge() operation of CodeOwnerConfigValidator was creating
a RepoView on-the-fly without a proper try-with-resource causing
the leak of a Repository object reference counting.

Extract the repoView creation in a separate variable and make sure
it is included in the outer try-with-resource everywhere it is used.

NOTE: ErrorProne should have spotted resources leaks in the
first place, so we should consider enabling the MustBeClosedChecker.

[1] https://errorprone.info/bugpattern/MustBeClosedChecker

Change-Id: I2880a4cf8b76e2be979288c996720478f98b4fba