Return code-owners~skip-validation option only if OWNERS files are touched
Skipping the code owners config validation is only relevant if the
change contains code owner config files (e.g. OWNERS files) that can be
validated. We should hide the skip option if for changes that do not
touch code owner config files.

Note, it depends on the configuration of the code-owners plugin for
which file names we need to check. This has to be taken into account
when we check whether code owner config files have been touched.

Bug: Google b/279897514
Change-Id: I99b816873dc75a74fec31da623ca27b79949c474
Signed-off-by: Edwin Kempin <ekempin@google.com>