Fix RepoView leak in CodeOwnerConfigValidator
The onRefOperation() operation of CodeOwnerConfigValidator
was creating a RepoView on-the-fly without a proper
try-with-resource causing the leak of a Repository object
reference counting.

Extract the RepoView creation in onRefOperation in a
separate variable.

NOTE: ErrorProne could have spotted resources leaks in the
first place, if RepoView was declared AutoCloseable and tagged
with @MustBeClosed. That could now be possible but it would
require a lot more invasive changes in JGit and Gerrit.

[1] https://errorprone.info/bugpattern/MustBeClosedChecker

Change-Id: I096453aed25c3d57bafdb8c8645233c04155adfd