Increate JVM Heap for the acceptance tests execution
The execution of CodeOwnerConfigValidatorIT requires a lot of
JVM Heap because of the continuous restarts of Gerrit daemon with
different configuration settings and the number of tests executed
against it.

The default 256MBytes is bearly enough for a test-run with multiple
GC cycles at the end; raising the limit to 512MBytes allows more
space for smooth executions without GCs and less flakiness due
to out of memory errors.

Change-Id: I9b10b3406fc46c60f4a645a909d34c730b462f49