ChangedFilesTest: Set project when creating test changes
When creating a test change without specifying a project the change is
created in an arbitrary project. Since the project is arbitrary tests
that create multiple changes cannot assume that all changes are created
in the same project, when no project is being specified.

So far this was not a problem, since always the same arbitrary project
has been selected, but with change I88f26dd6a in Gerrit core this is
changing now. That change makes it so that every change is created in a
new project, if no project was specified.

This makes some tests in ChangedFilesTest fail. Fix them by setting the
project when creating test changes.

Change-Id: Ic61a2bfc687a9081d666bc1524df7e2d41f7d9b5
Signed-off-by: Edwin Kempin <ekempin@google.com>