Drop CodeOwnersInternalServerErrorException
CodeOwnersInternalServerErrorException was added during the rollout of
the code-owners feature to recognise issues with the code-owners plugin
immediately. It's thrown for any exception that happens in the
code-owners plugin, even if the error is not caused by the code-owners
plugin, but happens in other Gerrit modules that are used by the
code-owners plugin. For example if permissions cannot be evaluated and
the PermissionBackend fails with a PermissionBackendException, anywhere
in Gerrit that would result in a normal "internal server error", only if
the caller is the code-owners plugin it is a "code-owners internal
server error". This doesn't make sense, at least no longer since the
code-owners rollout is long done. In retrospect wrapping arbitrary
exceptions that happen in code that is called from the code-owners
plugin into CodeOwnersInternalServerErrorException's caused more
confusion than it helped because non-code-owner related errors were
attributed to the code-owners plugin. Hence dropping this exception now.
Instead we just use StorageException like in the rest of the Gerrit
code.

Change-Id: I77e7e2026de7a6c3a6eccd184fd20993dc81aa22
Signed-off-by: Edwin Kempin <ekempin@google.com>