ChangedFiles: Rename 'getFromDiffCache' methods to 'get'
Earlier ChangedFiles offered 2 alternatives for geting the changed
files, one that computed the changed files and one that retrieved them
from the diff cache. The first alternative was dropped in change
I811c28f64 and now the changed files are always retrieved from the diff
cache. Back then when the 2 alternatives existed it was important that
they could be distinguished by the method, but now that only 1 way to
get the changed files exists the "FromDiffCache" part of the method
names that exposes an implementation detail is no longer needed.

Dropping the "FromDiffCache" part from the method names will allow us to
change the implementation later, e.g. to get the changed files in
another way than retrieving them from the diff cache.

Change-Id: I454e4896a7295432306c6bd620e805edf9ae245a
Signed-off-by: Edwin Kempin <ekempin@google.com>