Avoid computing file diffs when evalution code owner approvals
To evaluate code owner approvals on a change we need to know which files
have been changed. So far we retrieved the changed files from the diff
cache, which means file diffs are being computed when they are not
cached yet (or when their cache entry has expired). Computing files diff
for changes that touch many files (e.g. > 70k files) can take > 15min.
Since the code owner approval evaluation only needs to know which files
have changed, computing the file diffs seems unnecessary. DiffOperations
has a new method that allows to get the modified files without computing
the file diffs and with this change we are using it to get the changed
files instead of getting them from the diff cache. For the evalution of
code owner approvals detecting renames is important (because renamed
files require an approval on the old and the new path), hence the
rename detection must be enabled.

Bug: Google b/461456634
Change-Id: I45c7e61832be45676ef0ceae5044990cfdf469e0
Signed-off-by: Edwin Kempin <ekempin@google.com>