SkipCodeOwnerConfigValidationPushOption: Get changed files without rename detection
The code only needs to know if any OWNERS file has been added or
updated. For this is only looks at the new paths of the changed files.
Hence rename detection is not required here.

If the rename detection is enabled, renamed files are retured as a
single ChangedFile (with old and new path). If the rename detection is
disabled, renamed files are returned as two ChangeFiles (one with the
old path for the deletion and one with the new path for the addition).
Verify this behaviour in a test.

Change-Id: If42e8410045d38613e6292e8d6e0e87575214ab0
Signed-off-by: Edwin Kempin <ekempin@google.com>