ChangedFilesByPatchSetCache: Get changed files without rename detection
The only caller of ChangedFilesByPatchSetCache is
CodeOwnerApprovalCheck. CodeOwnerApprovalCheck only uses
ChangedFilesByPatchSetCache to check whether the path has been
code-owner approved in a previous patch set. For this it doesn't matter
whether renamed are represented as a single ChangedFile (with new and
old path set) or by 2 separate ChangedFile's (one that only has the new
path set and one that only has the old path set).

Bug: Google b/461456634
Change-Id: Ifdea3fed3379132d6271270e27866eb7a988f2ec
Signed-off-by: Edwin Kempin <ekempin@google.com>