CodeOwnersInChangeCollection: Get changed paths without rename detection
CodeOwnersInChangeCollection only needs to check whether the path exists
as a new path or as an old path of a rename or deletion. Since old paths
of deletions and renames are handled the same way we do not need to
detect renames. If we get the changed files without rename detection,
renames are returned as 2 changed files, one with the new path for the
addition and one with the old path for the deletion.

Change-Id: I9a857e0bf17262a7be56c7a2affcc2cbe4f12694
Signed-off-by: Edwin Kempin <ekempin@google.com>