Disable rename detection for the code owners submit rule
The code owners submit rule only checks if all paths (new and old) have
been code owner approved (see CodeOwnerApprovalCheck#isSubmittable). For
this it's OK if for renamed files 2 FileCodeOwnerStatus'es are returned
(one for the new path and one for the old path), hence rename detection
can be disabled here.

When code owner statuses are returned from the GetCodeOwnerStatus REST
endpoint however we must detect renames (see
CodeOwnerApprovalCheck#getFileStatusesAsSet) because our API
documentation for this REST endpoint says that we return a single code
owner status for renamed files. This is important for callers that do
pagination, e.g. if the change screen shows the first 200 files (which
may include renames) than getting the first 200 code owner statuses
should match these files (their new and old paths).

Changing CodeOwnerApprovalCheck#getAllPathsAsApproved to always disable
the rename detection didn't make any tests fail although it affects the
GetCodeOwnerStatus REST endpoint. This was because getAllPathsAsApproved
is only called from the GetCodeOwnerStatus REST endpoint if a change
contains a renamed file and all files of the change are exempted from
requiring code owner approvals because a) the uploader is exempted from
requiring code owner approvals or b) the change is a pure revert and
pure reverts are configured to not require code owner approvals. The new
tests in GetCodeOwnerStatusIT cover this now.

CodeOwnerApprovalCheck has one more caller of ChangedFiles for which we
haven't checked yet whether rename detection can be disabled. At this
place we add a TODO so that we remember to check this later.

Bug: Google b/461456634
Change-Id: I9fc7ca3a808a176324fc9eb2f1d49b4bb4b55cae
Signed-off-by: Edwin Kempin <ekempin@google.com>